Move csl files
