+++ include/csl/glsl/Types.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../include/Types.hpp"
+#include <csl/Types.hpp>
 
 namespace csl
 {
