+++ shader_suite/shaders/readme_examples.cpp
