+++ shader_suite/shaders/rendering.cpp
